+++ Debaser/Internals/Exceptions/EmptySequenceException.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Runtime.Serialization;
+
+namespace Debaser.Internals.Exceptions
+{
+    [Serializable]
+    class EmptySequenceException : Exception
+    {
+        public EmptySequenceException()
+        {
+        }
+
+        public EmptySequenceException(SerializationInfo info, StreamingContext context) : base(info, context)
+        {
+        }
+    }
+}